+++ docs/Whats-New.md
@@ -52,7 +52,7 @@ New:
 - XDrawOffset for animations (by Morton)
 - Customizable OpenTopped properties (by Otamaa)
 - Automatic Passenger Deletion (by FS-21)
-- Script Action 74 to 81 for new AI attacks (by FS-21)
+- Script Action 74 to 81 & 104 to 105 for new AI attacks (by FS-21)
 - Script Actions 82 & 83 for modifying AI Trigger Current Weight (by FS-21)
 - Script Action to regroup temporarily around the Team Leader (by FS-21)